Archive to a directory
- Store the results of the run into a CSV.
- Minor cleanup
